Made the timer .1 sec faster on wins.

+++ Header/InaTransitionTimer.h
@@ -6,7 +6,7 @@
 class InaTransitionTimer : public GameObject
 {
 public:
-    InaTransitionTimer(sf::Vector2f position, int sec, Camera* camera, Scene* scene, float scale = 1.0f);
+    InaTransitionTimer(sf::Vector2f position, float sec, Camera* camera, Scene* scene, float scale = 1.0f);
     ~InaTransitionTimer();
     void Update(float delta_time) override;
     void Draw() override;
@@ -14,7 +14,7 @@ public:
 private:
     Camera* camera_{nullptr};
     float scale_{1.0f};
-    int sec_{0};
+    float sec_{0};
     float timer_{1.0f};
     int end_{0};
     int bomb_{0};
